+++ nautobot/ui/src/constants/icons.js
@@ -1,0 +1,15 @@
+import {
+    AutomationIcon,
+    DcimIcon,
+    IpamIcon,
+    PlatformIcon,
+    SecurityIcon,
+} from "@nautobot/nautobot-ui";
+
+export const appContextIcons = {
+    Inventory: DcimIcon,
+    Networks: IpamIcon,
+    Security: SecurityIcon,
+    Automation: AutomationIcon,
+    Platform: PlatformIcon,
+};
