+++ nautobot/ui/src/utils/store.js
@@ -33,15 +33,92 @@ const appContextSlice = createSlice({
     },
 });
 
+const initialAppState = {
+    currentContext: "Inventory", // TODO: What is the context for the homepage?
+    routeToContext: {},
+};
+
+const appStateSlice = createSlice({
+    name: "appState",
+    initialState: initialAppState,
+    reducers: {
+        updateAppCurrentContext(state, action) {
+            state.currentContext = action.payload;
+            return state;
+        },
+        updateRouteToContext(state, action) {
+            let urlPatternToContext = {};
+            let menuInfo = action.payload;
+            for (const context in menuInfo) {
+                for (const group in menuInfo[context].groups) {
+                    for (const urlPatternOrSubGroup in menuInfo[context].groups[
+                        group
+                    ].items) {
+                        if (urlPatternOrSubGroup.startsWith("/")) {
+                            // It's a URL pattern
+                            let tokens = urlPatternOrSubGroup.split("/");
+                            if (tokens.length === 4) {
+                                let appLabel = tokens[1];
+                                let modelNamePlural = tokens[2];
+                                if (appLabel in urlPatternToContext === false) {
+                                    urlPatternToContext[appLabel] = {};
+                                }
+                                urlPatternToContext[appLabel][modelNamePlural] =
+                                    context;
+                            }
+                        } else {
+                            // It's a submenu
+                            const subGroup = urlPatternOrSubGroup;
+                            for (const urlPattern in menuInfo[context].groups[
+                                group
+                            ].items[subGroup].items) {
+                                let tokens = urlPattern.split("/");
+                                if (tokens.length === 4) {
+                                    let appLabel = tokens[1];
+                                    let modelNamePlural = tokens[2];
+                                    if (
+                                        appLabel in urlPatternToContext ===
+                                        false
+                                    ) {
+                                        urlPatternToContext[appLabel] = {};
+                                    }
+                                    urlPatternToContext[appLabel][
+                                        modelNamePlural
+                                    ] = context;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            state.routeToContext = urlPatternToContext;
+            return state;
+        },
+    },
+});
+
+export function getCurrentAppContextSelector(app_label, model_name) {
+    return (store) => {
+        return (
+            store?.appState?.routeToContext[app_label][model_name] ||
+            "Inventory"
+        );
+    };
+}
+
 const rootReducer = combineReducers({
     [baseApi.reducerPath]: baseApi.reducer,
     appContext: appContextSlice.reducer,
+    appState: appStateSlice.reducer,
 });
 
 // Instantiate the persistent reducer to pull state from cache
 const persistedReducer = persistReducer(persistConfig, rootReducer);
 
 export const { updateAppContext } = appContextSlice.actions;
+export const { updateAppCurrentContext, updateRouteToContext } =
+    appStateSlice.actions;
 
 // Global Redux store for global state management
 export const store = configureStore({
