+++ nautobot/ui/src/views/generic/GenericView.js
@@ -1,34 +1,14 @@
 import {
-    AutomationIcon,
     Box,
     Breadcrumb,
     Breadcrumbs,
-    DcimIcon,
     Flex,
-    Input,
-    InputGroup,
-    InputLeftElement,
-    IpamIcon,
-    Menu,
-    MenuButton,
-    MenuItem,
-    MenuList,
     NautobotGrid,
-    Navbar,
-    NavbarMenuButton,
-    NavbarSection,
-    NavbarSections,
-    PlatformIcon,
-    SearchIcon,
-    SecurityIcon,
 } from "@nautobot/nautobot-ui";
+import { Navbar } from "@components/Navbar";
 import { useMemo } from "react";
 import { useSelector } from "react-redux";
-import {
-    Link as ReactRouterLink,
-    NavLink as ReactRouterNavLink,
-    useLocation,
-} from "react-router-dom";
+import { Link as ReactRouterLink, useLocation } from "react-router-dom";
 
 import { useGetSessionQuery, useGetUIMenuQuery } from "@utils/api";
 
@@ -39,7 +19,6 @@ export default function GenericView({
     rows,
 }) {
     const { pathname } = useLocation();
-
     const menu = useGetUIMenuQuery();
     const session = useGetSessionQuery();
 
@@ -232,9 +211,9 @@ export default function GenericView({
         [menu, objectData, pathname]
     );
 
-    const isLoggedIn = !!session?.data?.logged_in;
-
-    const appContext = useSelector((state) => state.appContext);
+    const currentContext = useSelector(
+        (state) => state.appState.currentContext
+    );
 
     return (
         <Flex
@@ -245,93 +224,7 @@ export default function GenericView({
             paddingTop="md"
             width="full"
         >
-            <Navbar>
-                <NavbarSections>
-                    {[
-                        {
-                            children: "Inventory",
-                            leftIcon: <DcimIcon />,
-                            to: "/dcim/devices/",
-                        },
-                        {
-                            children: "Networks",
-                            leftIcon: <IpamIcon />,
-                            to: "/ipam/ip-addresses/",
-                        },
-                        {
-                            children: "Security",
-                            leftIcon: <SecurityIcon />,
-                            to: "/security",
-                        },
-                        {
-                            children: "Automation",
-                            leftIcon: <AutomationIcon />,
-                            to: "/extras/jobs/",
-                        },
-                        {
-                            children: "Platform",
-                            leftIcon: <PlatformIcon />,
-                            to: "/extras/relationships/",
-                        },
-                    ].map(({ children, to, ...rest }) => (
-                        <ReactRouterNavLink key={to} to={to}>
-                            {({ isActive }) => (
-                                <NavbarSection
-                                    as="span"
-                                    isActive={
-                                        isActive || children === appContext
-                                    }
-                                    children={children}
-                                    {...rest}
-                                />
-                            )}
-                        </ReactRouterNavLink>
-                    ))}
-                </NavbarSections>
-
-                <InputGroup flex="1" size="lg" variant="navbar">
-                    <InputLeftElement>
-                        <SearchIcon />
-                    </InputLeftElement>
-                    <Input placeholder="Search..." />
-                </InputGroup>
-
-                <Menu>
-                    <MenuButton as={NavbarMenuButton} isDisabled={!isLoggedIn}>
-                        {session?.data?.user?.display ||
-                            [
-                                ...(session?.data?.user?.firstName
-                                    ? [session?.data?.user?.firstName]
-                                    : []),
-                                ...(session?.data?.user?.lastName
-                                    ? [session?.data?.user?.lastName]
-                                    : []),
-                            ].join(" ") ||
-                            session?.data?.user?.username}
-                    </MenuButton>
-                    <MenuList>
-                        {[
-                            {
-                                children: "Profile",
-                                to: "/user/profile",
-                            },
-                            {
-                                children: "Log Out",
-                                color: "red-1",
-                                to: "/logout",
-                            },
-                        ].map(({ to, ...rest }) => (
-                            <MenuItem
-                                key={to}
-                                as={ReactRouterLink}
-                                to={to}
-                                {...rest}
-                            />
-                        ))}
-                    </MenuList>
-                </Menu>
-            </Navbar>
-
+            <Navbar session={session?.data} currentContext={currentContext} />
             <Box flex="1" overflow="auto">
                 <Breadcrumbs paddingX="md">
                     {breadcrumbs.map((props) => (
