+++ nautobot/ui/src/views/generic/ObjectList.js
@@ -5,11 +5,23 @@ import { useDispatch } from "react-redux";
 import { LoadingWidget } from "@components/LoadingWidget";
 import ObjectListTable from "@components/ObjectListTable";
 import GenericView from "@views/generic/GenericView";
-import { useGetUIMenuQuery, useGetRESTAPIQuery } from "@utils/api";
-import { updateAppContext } from "@utils/store";
+import { useGetRESTAPIQuery } from "@utils/api";
+import { useEffect } from "react";
+import {
+    updateAppCurrentContext,
+    getCurrentAppContextSelector,
+} from "@utils/store";
+import { useSelector } from "react-redux";
 
 export default function GenericObjectListView() {
     const { app_name, model_name } = useParams();
+    const dispatch = useDispatch();
+    const currentAppContext = useSelector(
+        getCurrentAppContextSelector(app_name, model_name)
+    );
+    useEffect(() => {
+        dispatch(updateAppCurrentContext(currentAppContext));
+    }, [dispatch, currentAppContext]);
     // const { 0: searchParams } = useSearchParams(); // import { useSearchParams } from "react-router-dom";
     const { data: listData, isLoading: listDataLoading } = useGetRESTAPIQuery({
         app_name: app_name,
@@ -21,55 +33,6 @@ export default function GenericObjectListView() {
             model_name: model_name,
             schema: true,
         });
-
-    const {
-        data: menuInfo,
-        isSuccess: isMenuSuccess,
-        isError: isMenuError,
-    } = useGetUIMenuQuery();
-    const dispatch = useDispatch();
-
-    if (isMenuError) return <div>Failed to load menu</div>;
-    if (!isMenuSuccess) return <span>Loading...</span>;
-
-    // Construct reverse-lookup from URL patterns to menu context
-    var urlPatternToContext = {};
-    for (const context in menuInfo) {
-        for (const group in menuInfo[context].groups) {
-            for (const urlPatternOrSubGroup in menuInfo[context].groups[group]
-                .items) {
-                if (urlPatternOrSubGroup.startsWith("/")) {
-                    // It's a URL pattern
-                    let tokens = urlPatternOrSubGroup.split("/");
-                    if (tokens.length === 4) {
-                        let appLabel = tokens[1];
-                        let modelNamePlural = tokens[2];
-                        if (appLabel in urlPatternToContext === false) {
-                            urlPatternToContext[appLabel] = {};
-                        }
-                        urlPatternToContext[appLabel][modelNamePlural] =
-                            context;
-                    }
-                } else {
-                    // It's a submenu
-                    const subGroup = urlPatternOrSubGroup;
-                    for (const urlPattern in menuInfo[context].groups[group]
-                        .items[subGroup].items) {
-                        let tokens = urlPattern.split("/");
-                        if (tokens.length === 4) {
-                            let appLabel = tokens[1];
-                            let modelNamePlural = tokens[2];
-                            if (appLabel in urlPatternToContext === false) {
-                                urlPatternToContext[appLabel] = {};
-                            }
-                            urlPatternToContext[appLabel][modelNamePlural] =
-                                context;
-                        }
-                    }
-                }
-            }
-        }
-    }
 
     // What page are we on?
     // TODO: Pagination handling should be it's own function so it's testable
@@ -91,8 +54,6 @@ export default function GenericObjectListView() {
             </GenericView>
         );
     }
-
-    dispatch(updateAppContext(urlPatternToContext[app_name][model_name]));
 
     if (listDataLoading || headerDataLoading) {
         return (
